Pass params to Faraday correctly in the PurlFetcher::Client::Reader
Fixes #61

This passes the parameters to append to the request to Faraday
in the correct way – as a second parameter, rather than a hash with
key 'params'.

The latter approach resulted in the params being rendered at the
literal key 'params', which encoded and collapsed all of them
inside the URL, and as a result purl-fetcher did not interpret them.

This in turn caused pagination to not work because we were always
requesting the same page of results over and over again.

The presence of a test that stubs the params in this way seems to
suggest that purl-fetcher's API once might have worked this way,
but it no longer does.